Move deprecated functions into separate module
These functions were causing cyclical import problems.
